refactor: begin to refactor
