+++ lib/services/__tests__/auth.test.js
@@ -1,16 +1,33 @@
-import {
-  initConnection,
-  createPermissionResult,
-  approveConnection,
-} from '../auth'
+import { initConnection, approveConnection } from '../auth'
 import { getAccount } from '../account'
 import { generateTestAccount } from './_helpers'
 import { generateKey } from '../crypto'
 import Config from 'react-native-config' // this is mocked
 import * as operatorAdapter from '../operatorAdapter'
 
+jest.unmock('@egendata/react-native-jose')
+jest.mock('../jwt.js', () => {
+  const realReactNativeJose = jest.requireActual('react-native-jose')
+  return {
+    verify: realReactNativeJose.verify,
+    // sign: jest
+    //   .fn()
+    //   .mockResolvedValue()
+    //   .mockResolvedValue('jwt-for-createConnectionResponse'),
+    // verify: jest.fn().mockResolvedValue({ payload: { foo: 'bar' } }),
+  }
+})
 jest.mock('../account', () => ({
   getAccount: jest.fn(),
+}))
+jest.mock('../storage', () => ({
+  getAccountKeys: jest.fn().mockReturnValue({
+    publicKey: 'pub key',
+    privateKey: 'priv key',
+    privateKeyPem: 'priv key pem',
+  }),
+  getAccount: jest.fn().mockReturnValue({ id: 1 }),
+  storeConnection: jest.fn(),
 }))
 
 jest.mock('../serviceAdapter', () => ({
@@ -18,16 +35,18 @@ jest.mock('../serviceAdapter', () => ({
     .fn()
     .mockResolvedValue('jwt-for-createConnectionInit'),
 }))
-jest.mock('../operatorAdapter', () => ({
-  createConnection: jest.fn().mockResolvedValue('jwt-for-createConnection'),
-  createConnectionResponse: jest
-    .fn()
-    .mockResolvedValue('jwt-for-createConnectionResponse'),
-}))
-
-jest.mock('../jwt', () => ({
-  verify: jest.fn().mockResolvedValue({ payload: { foo: 'bar' } }),
-}))
+jest.mock('../operatorAdapter', () => {
+  const operatorAdapterActual = jest.requireActual('../operatorAdapter.js')
+  return {
+    createConnection: jest.fn().mockResolvedValue('jwt-for-createConnection'),
+    getApprovedPermissionRequestWithKeys:
+      operatorAdapterActual.getApprovedPermissionRequestWithKeys,
+    newCreateConnection: operatorAdapterActual.newCreateConnection,
+    createConnectionResponse: jest
+      .fn()
+      .mockResolvedValue('jwt-for-createConnectionResponse'),
+  }
+})
 
 jest.mock('../crypto', () => {
   const actualCrypto = jest.requireActual('../crypto')
@@ -175,7 +194,7 @@ describe('auth', () => {
         ],
       }
     })
-    describe('#createPermissionResult', () => {
+    describe('#operatorAdapter.getApprovedPermissionRequestWithKeys', () => {
       beforeEach(() => {
         generateKey.mockResolvedValueOnce({ privateKey: appKeyBaseData })
         generateKey.mockResolvedValueOnce({ privateKey: appKeyExperience })
@@ -245,7 +264,10 @@ describe('auth', () => {
           ['fc284cf5-b1af-4fac-b793-7d1adf8a9c60', true],
         ])
 
-        const result = await createPermissionResult(connectionRequest, approved)
+        const result = await operatorAdapter.getApprovedPermissionRequestWithKeys(
+          connectionRequest,
+          approved,
+        )
 
         expect(result).toEqual(expected)
       })
@@ -260,7 +282,6 @@ describe('auth', () => {
       })
       it('calls operatorAdapter.createConnection with correct arguments for empty approved', async () => {
         await approveConnection(connectionRequest, new Map())
-
         expect(operatorAdapter.createConnection).toHaveBeenCalledWith(
           connectionRequest,
           undefined,
