+++ lib/services/operatorAdapter.js
@@ -1,7 +1,9 @@
 import { sign } from './jwt'
-import { getAccount, getAccountKeys, getConnection } from './storage'
+import { getAccount, getAccountKeys, getConnection, storeKey } from './storage'
 import Config from 'react-native-config'
 import { schemas } from '@egendata/messaging'
+import { v4 } from 'uuid'
+import { generateKey, toPublicKey } from './crypto'
 
 export const createAccountRegistration = async (
   // eslint-disable-next-line camelcase
@@ -135,4 +137,91 @@ export function postToOperator(body) {
     headers: { 'content-type': 'application/jwt' },
     body,
   })
+}
+
+export async function newCreateConnection(
+  connectionRequest,
+  approvedPermissions,
+) {
+  const connectionId = v4()
+
+  const permissionsResult = await getApprovedPermissionRequestWithKeys(
+    connectionRequest,
+    approvedPermissions,
+  )
+  const connection = await createConnection(
+    connectionRequest,
+    permissionsResult,
+    connectionId,
+  )
+
+  const connectionResponse = await createConnectionResponse(connection)
+
+  await fetch(Config.OPERATOR_URL, {
+    method: 'POST',
+    headers: { 'content-type': 'application/jwt' },
+    body: connectionResponse,
+  })
+
+  return connectionId
+}
+
+export async function getApprovedPermissionRequestWithKeys(
+  { permissions: requestedPermissions },
+  approved,
+) {
+  if (!requestedPermissions) {
+    return undefined
+  }
+
+  const readServiceReadKeysByArea = mapReadKeys(requestedPermissions)
+
+  const permissionResult = {
+    approved: await Promise.all(
+      requestedPermissions
+        .filter(p => approved.get(p.id))
+        .map(async p => {
+          if (p.type === 'WRITE') {
+            if (!p.jwks) {
+              p.jwks = {
+                keys: [],
+              }
+            }
+            // push service read-keys to jwks-keys
+            const serviceReadKey = readServiceReadKeysByArea.get(
+              `${p.domain}|${p.area}`,
+            )
+            if (serviceReadKey) {
+              p.jwks.keys.push(serviceReadKey)
+            }
+
+            // push user read-key
+            const key = await generateKey({
+              use: 'enc',
+            }).then(storeKey)
+
+            p.jwks.keys.push(toPublicKey(key.privateKey))
+          } else if (p.type === 'READ') {
+            // todo: l8r if there are no recipients then create JWE with empty data
+
+            // todo: get jwe recipients
+
+            // if there are then add it to the recipeintseassesez
+
+            p.kid = p.jwk.kid
+            delete p.jwk
+          }
+          return p
+        }),
+    ),
+  }
+  return permissionResult.approved.length ? permissionResult : undefined
+}
+
+function mapReadKeys(permissions) {
+  return permissions
+    .filter(p => p.type === 'READ')
+    .reduce((map, { domain, area, jwk }) => {
+      return map.set(`${domain}|${area}`, jwk)
+    }, new Map())
 }
